Add tempo view and working main controls
Moved some responsibilities around so that they make slightly more sense
to me. I'm not happy with my current ENUM workaround, especially since
I'm using two 'different' systems in the same place. This means that one
can be mistaken for the other without a bunch of stupid flags.

I should also spend some time in moving the update delta logic to the
LaunchPad controller instead of keeping it in the tempo view. It makes
more sense to keep track of the state of the entire LP and diff there
instead of doing it in all places seperately.

I'd also like to update the buttons abstraction to be universal instead
of having variations on it across the different components.

Also, should reevaluate the pad abstraction and maybe rename the value,
since in this case it can only be ON or OFF anyway so it makes sense to
just use a boolean.